chore: 불필요한 함수 삭제 및 validate 함수 파일 이동
