+++ frontend/src/shared/utils/convertTime.ts
@@ -11,14 +11,6 @@ export const secondsToMinSec = (seconds: number): TimeMinSec => {
   };
 };
 
-export const calculateMinSec = (
-  minute: number,
-  second: number,
-  calculateFn: (originalSecond: number) => number
-) => {
-  return secondsToMinSec(calculateFn(minSecToSeconds({ minute, second })));
-};
-
 export const toMinSecText = (seconds: number) => {
   const { minute, second } = secondsToMinSec(seconds);
 
@@ -30,24 +22,4 @@ export const toMinSecText = (seconds: number) => {
 
 export const toPlayingTimeText = (startTime: number, endTime: number) => {
   return `${toMinSecText(startTime)} ~ ${toMinSecText(endTime)}`;
-};
-
-export const covertTwoDigitTimeFormat = (content: string | number) => {
-  if (typeof content === 'number') {
-    return content.toString().padStart(2, '0');
-  }
-
-  return content.padStart(2, '0');
-};
-
-export const covertTwoDigitTimeFormatText = (...contents: (string | number)[]) => {
-  return contents.map(covertTwoDigitTimeFormat).join(':');
-};
-
-export const isValidTimeFormatLength = (value: string | number) => {
-  if (typeof value === 'number') {
-    return value.toString().length <= 2;
-  }
-
-  return value.length <= 2;
 };
